Дипломная работа
Исправлены ошибки.
Доработан файл README.

+++ controllers/movie.js
@@ -1,6 +1,7 @@
 const Movie = require('../models/movie');
 const ValidationError = require('../errors/ValidationError');
 const NotFoundError = require('../errors/NotFoundError');
+const ForbiddenError = require('../errors/ForbiddenError');
 
 function getSavedMovies(req, res, next) {
   Movie.find({ owner: req.user._id })
@@ -43,9 +44,12 @@ function deleteMovie(req, res, next) {
     .then((movie) => {
       if (!movie) {
         throw new NotFoundError('Запрашиваемый фильм не найден');
-      } else {
+      }
+      if (String(movie.owner) === req.user._id) {
         Movie.findByIdAndRemove(req.params._id)
           .then(() => res.send(movie));
+      } else {
+        throw new ForbiddenError('Запрещено удалять чужие сохраненные фильмы');
       }
     })
     .catch(next);
